Went Sicko Mode and redid the start button so it isn't using tkinker just pygame.
it doesn't work but google and such say it's because the raycasting or something I haven't worked on. have been working on enemies but not on this repository so you won't see anything regarding that.

also i had to update from 3.12 to 3.9 but I thought I already did that.
nick woerner